Added optbypairwiseiter, an algorithm to find optimal portfolios.
In investor.create, added pretax.pctwts to the investor. Also added totalassets.at

In rafi.cma, reduced dependency on the Get10YrBEInflationRate function which requires Quandl and may have problems if many calls are made and the user is not using a Quandl key.

In the cma.ta.create function, created the boxmin and boxmax to be relevant after-tax (0 for min, and the smaller of the pre-tax boxmax and the account size for the after-tax boxmax.